Grants table now displays only grants for the currently logged in user
