+++ app/templates/grants/index.hbs
@@ -2,7 +2,7 @@
   <div class="col-sm-12">
         <h1 class="font-weight-light">Grants</h1>
         {{models-table
-          data=tableModel
+          data=model
           columns=columns
           themeInstance=themeInstance
           showColumnsDropdown=false
